feat(stracchino-22301): cap EventPage "On the Menu" to top-3 pizzerias
Lifts rankPizzerias() out of useRSVPForm.ts into a shared
frontend/src/lib/pizzeriaRank.ts util and applies it to the public
EventPage "On the Menu" section. RSVP form behavior unchanged — both
surfaces now show the same top-3 in the same order, ranked by
(rating - distance*0.3) just like vesuvio-58492 introduced.
