+++ frontend/src/hooks/useRSVPForm.ts
@@ -1,43 +1,14 @@
 import { useState, useEffect, useCallback } from 'react';
 import { addGuestToParty, getUserPreferences, saveUserPreferences, ExistingGuestData, getExperimentFlag } from '../lib/supabase';
 import { getExcludedToppingIds } from '../constants/options';
-import { searchPizzerias, geocodeAddress, calculateDistanceMiles } from '../lib/ordering';
+import { searchPizzerias, geocodeAddress } from '../lib/ordering';
 import { Pizzeria } from '../types';
 import { PublicEvent, trackRsvpFunnel } from '../lib/api';
 import { DbParty } from '../lib/supabase';
 import { uuid } from '../lib/utils';
 import { findActiveRegion } from '../lib/optinAbRegions';
 import { getOrCreateVisitorSessionId } from '../lib/visitorSession';
-
-// ---- Ranking helpers (vesuvio-58492) ----
-
-// Cap the RSVP "Favorite Pizzerias" list to the top N entries ranked by a
-// weighted score combining rating and distance from the venue. Istanbul GPP
-// had ~15+ host-selected pizzerias which overwhelmed the form.
-const TOP_PIZZERIA_LIMIT = 3;
-const DISTANCE_WEIGHT_PER_MILE = 0.3;
-
-function rankPizzerias(
-  list: Pizzeria[],
-  venue: { lat: number; lng: number } | null,
-): Pizzeria[] {
-  return [...list]
-    .map(p => {
-      const rating = p.rating ?? 3.5;
-      const hasDistance =
-        !!venue &&
-        !!p.location &&
-        p.location.lat !== 0 &&
-        p.location.lng !== 0;
-      const distance = hasDistance
-        ? calculateDistanceMiles(venue.lat, venue.lng, p.location.lat, p.location.lng)
-        : 0;
-      return { p, score: rating - distance * DISTANCE_WEIGHT_PER_MILE };
-    })
-    .sort((a, b) => b.score - a.score)
-    .slice(0, TOP_PIZZERIA_LIMIT)
-    .map(x => x.p);
-}
+import { rankPizzerias } from '../lib/pizzeriaRank';
 
 // ---- Types ----
